% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{boxplot2}
\alias{boxplot2}
\title{Boxplot for function 1}
\usage{
boxplot2(data, grouping_column, payment_column)
}
\arguments{
\item{data}{the dataset}

\item{grouping_column}{the grouping variable}

\item{payment_column}{payment type}
}
\value{
A boxplot between DRG Code(Grouping variable) and payment type
}
\description{
Boxplot for function 1
}
\examples{
boxplot2(DRG,'DRG.Definition', 'Average.Covered.Charges')

}
